+++ src/server/orderBook.c
@@ -3,13 +3,13 @@
 void addToBook(order * curOrder, book ** curBook){
 	book * tempBook = NULL;
 	priceBucket * tempPb = NULL;
-	float tempPrice = curOrder -> price;
+	double tempPrice = curOrder -> price;
 	/*
 	* check if the price is already in the book
 	* if not, create a new price bucket and add it to book
 	* else, add the order to corresponding price bucket retrieval from book
 	*/
-	HASH_FIND(hh, *curBook, &tempPrice, sizeof(float), tempBook);
+	HASH_FIND(hh, *curBook, &tempPrice, sizeof(double), tempBook);
 	if(tempBook == NULL){
 		tempBook = malloc(sizeof(book));
 		tempBook -> price = tempPrice;
@@ -18,7 +18,7 @@ void addToBook(order * curOrder, book ** curBook){
 		tempPb -> exchangeId = (curOrder -> exchangeId) ;
 		tempPb -> curOrder = curOrder;
 		HASH_ADD_INT((tempBook -> pb), exchangeId, tempPb);
-		HASH_ADD(hh, *curBook, price, sizeof(float), tempBook);
+		HASH_ADD(hh, *curBook, price, sizeof(double), tempBook);
 	} else{
 		tempPb = malloc(sizeof(priceBucket));
 		tempPb -> exchangeId = (curOrder -> exchangeId);
@@ -30,9 +30,9 @@ void addToBook(order * curOrder, book ** curBook){
 void deleteFromBook(order * curOrder, book ** curBook){
     book * tempBook = NULL;
     priceBucket * tempPb = NULL;
-    float tempPrice = curOrder -> price;
+    double tempPrice = curOrder -> price;
     // check if the price of order exists in book
-    HASH_FIND(hh, *curBook, &tempPrice, sizeof(float), tempBook);
+    HASH_FIND(hh, *curBook, &tempPrice, sizeof(double), tempBook);
     if(tempBook == NULL){
         printf("not found in book \n");
         return;
